test_ui: fix flaky unittest involving sigint signal
When run in non-blocking mode, ui.Run installs a signal handler
and then starts a UI thread, which will raise the SIGINT signal.

It turns out that this signal may be caught before the Python
standard threading.Thread object has returned from its run()
function.

See the following code for proof of this:

  import os
  import signal
  import threading

  def SignalHandler(signum, frame):
    del signum, frame
    raise Exception

  def Func():
    os.kill(os.getpid(), signal.SIGINT)

  signal.signal(signal.SIGINT, SignalHandler)
  t = threading.Thread(target=Func)
  t.daemon = True

  try:
    t.start()
  except Exception:
    print('Exception thrown while thread starting!')

  try:
    t.join()
  except Exception:
    print('Exception thrown after thread starting!')

On my machine, "after" is printed for about 10% of runs.

The fix is to modify the unittest to join on the UI thread,
and wrap the assertRaises around both the ui.Run call as well
as a thread join() call.

BUG=None
TEST=Manually on local machine

Change-Id: I27101145c732005bb0c94931eed96acc6c4116d1
Reviewed-on: https://chromium-review.googlesource.com/420145
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>