instalog: implement flow policy to define allow/deny rules
Plugins can now be configured with a flow policy, which defines
a set of allow/deny rules.  Only if an Event matches against
these rules will it be sent to the plugin for processing.

The format of the configuration YAML file is also updated.

Old format:

  time:
    plugin: 'input_time'
    interval: 10

New format:

  time:
    plugin: 'input_time'
    targets: 'target_plugin_id'
    allow: []
    deny: []
    enable_recursion: False
    args:
      interval: 10

BUG=b:33736731
TEST=Manually on local machine

Change-Id: I2b4b897d05b998c82e0c37bcc710b6c2a866620c
Reviewed-on: https://chromium-review.googlesource.com/422116
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>