pytest: implement station-based Bluetooth test.
In this CL, we implement a station-based Bluetooth test. DUT would
scan the host station and try to connect with it. We assume the host
station can already act as a Bluetooth device, so the station setup
is easy, no need to prepare another Bluetooth device.

BUG=none
TEST=test at the device.

Change-Id: Ic820d03cf1cea9885b3b8d42739c1d22ca29c56d
Reviewed-on: https://chromium-review.googlesource.com/424272
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>