fix temporary file not deleted in time_sanitizer_unittest
We're actually only using the name of NamedTemporaryFile(). (The file is
deleted after that line), and didn't delete the file after we reuse it.

Change the code to override unittest.TestCase.run(), and use context
manager to manage the temporary folder, so it is automatically deleted
on exit.

BUG=chromium:665306
TEST=make test
TEST=in cros_sdk, rm -rf /tmp/tmp*, and run
     py/tools/time_sanitizer_unittest.py, make sure that no temp file
     /tmp/tmp* are left after the test complete.

Change-Id: Iba6ffbed0207c8518c5143a4ab200883589a5da3
Reviewed-on: https://chromium-review.googlesource.com/422764
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>