speed up test_factory_flow_unittest.
For BuildBoard, getting property arch needs some expensive process
calls, but the property may not be used at all. Make the property not
been set in initializer, but as a type_utils.LazyProperty, so the
expensive operation can be avoided if it's not used at all.

One of the place that doesn't use arch is test_factory_flow_unittest
(only used full_name), and the running time is reduced from 8.5 seconds
to 2.6 seconds.

BUG=chromium:665306
TEST=make test

Change-Id: I45fb96cdaa2789caba5954d429c08cfca63935fd
Reviewed-on: https://chromium-review.googlesource.com/423299
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>