init: Allow executing extra commands when inhibited jobs was fired.
Upstart is a based on state machine so we sometimes need to do few things, for
example firing Upstart events, when a job was inhibited.

For example, if job A depends on job B and we only want to inhibit job B (i.e.,
job  A must be executed), then we have to send 'started JOB=B' for Upstart to
continue.

To support this, common.d/inhibited_jobs/* now supports shell scripts so we can
send extra events, or do something else when a job was inhibited.

BUG=chromium:677915
TEST=Created 'trunksd' with 'initctl -n emit started JOB=trunksd' and see
     tpm_managerd was invoked.

Change-Id: Ic8f725660b7425d1a0de62299ced68691937456a
Reviewed-on: https://chromium-review.googlesource.com/424458
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>