device: fix all_serial_numbers in dut.info
Shouldn't cache all_serial_numbers or the values in it will be None most
of the times.

For station mode, SMT prefers to start without any cached data (serial
number) while FAT usually needs to load cached data, so we need an
argument to specify if DUT info should be loaded or not.

BUG=none
TEST=manually

Change-Id: I9c8a620e49dd5451612d4bdaf3706bb7ec4053eb
Reviewed-on: https://chromium-review.googlesource.com/425055
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>