testlog: modify behaviour when corrupt json is encountered
Previously, testlog's SeqGenerator would simply give up and raise
an exception when the last line of JSON was corrupt.

Now, it is updated to read every line and look for the maximum SEQ.
If none of the lines contain a valid SEQ, or if the file couldn't
be opened properly, use the system's time in milliseconds as the
initial SEQ value.

BUG=b:34064975
TEST=testlog_seq_unittest.py

Change-Id: I16057043695d498ca881229aa503feb9339a944b
Reviewed-on: https://chromium-review.googlesource.com/425576
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>