toybox: support overriding provider of each command
If toybox is compiled with different flags, some commands might not be
provided by toybox.  In this case, we might be able to replace the
command by the default one (in $PATH), or by other *box commands.

This CL provides an easy way to override providers of all or specific
commands.

For example:

  class XXBoard(chromeos.ChromeOSBoard):
    @component.DeviceProperty
    def toybox(self):
      # use dd from $PATH, df from busybox, and all other commands from
      # toolbox
      return toybox.Toybox(
          self, provider_map={'*': 'toolbox',
                              'dd': None,
                              'df': 'busybox'})

BUG=None
TEST=make test

Change-Id: I16bc6d15fba3b4d8476ec973e9d867bb0ce36a05
Reviewed-on: https://chromium-review.googlesource.com/425473
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>