hwid: add the new component which is not matched.
Originally while we add the component, we check the value not equal
to any of existed one. Since the value could be a regular expression
or the subset of the probed value, now we check the value not matched
with any of existed one.

BUG=none
TEST=hwid/v3/builder_unittest.py

Change-Id: I1d737be9927f46e59e8256f64f5d16fea77056c3
Reviewed-on: https://chromium-review.googlesource.com/426278
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>