setup: Add BOARD for cros_docker.
`cros_docker.sh umpire` uses `umpire` as container name, but when being
used under Dome, the name is `umpire_${board}`. To override that, you
have to specify a UMPIRE_CONTAINER_NAME variable, which is pretty long.

This change adds a BOARD environment variable that will change default
UMPIRE_CONTAINER_NAME to `umpire_${BOARD}` if BOARD is specified,
otherwise `umpire`.

When BOARD is not specified, it defaults to
${HOST_UMPIRE_DIR}/.default_board. Dome will update the file when a new
container was added.

BUG=None
TEST=./cros_docker.sh umpire shell
     UMPIRE_CONTAINER_NAME=umpire_eve ./cros_docker.sh umpire shell
     BOARD=eve ./cros_docker.sh umpire shell
     echo "eve" >/cros_umpire/.default_board && \
       ./cros_docker.sh umpire shell

Change-Id: If2c69b2ad5443035eba5d61e8f94624e80d14849
Reviewed-on: https://chromium-review.googlesource.com/426460
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>