goofy: instalog plugin: fix timeout parameter
If Goofy calls Testlog's Flush function without a timeout
value, the Flush will fail with an error like this:

  usage: cli.py flush [-h] [--timeout TIMEOUT] [plugin_id]
  cli.py flush: error: argument --timeout/-w: invalid float value: 'None'
  FlushException: Flush: Failed to flush Instalog output plugin:

This is due to chaining =None default arguments:
  Flush(timeout=None)
  ...calls...
  FlushInput(timeout=_DEFAULT_FLUSH_TIMEOUT)

Since the timeout=None value is passed onto FlushInput, the default
argument is not taken.  None is sent to Instalog's CLI as the timeout
value, and results in failure.

BUG=b:34065361
TEST=Manually on local machine

Change-Id: Ibbdd0d599870c3ab5614e001087c9faa6b0ddfc7
Reviewed-on: https://chromium-review.googlesource.com/426523
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>