raiden_cc_flip_check: Allow to start from any polarity.
Currently when ask_flip_operation is True, the test will check whether current
CC pin is CC1 or CC2 which is set as original_enabled_cc. What if we don't care
what CC pin now is and just would like to flip the calbe. The new value - None
of orignal_enabled_cc is added for this purpose.

BUG=chromium:680452
TEST=manual

Change-Id: I637e1f943cddfe173dbb51f303cff30b82796ceb
Reviewed-on: https://chromium-review.googlesource.com/428533
Commit-Ready: Marco Chen <marcochen@chromium.org>
Tested-by: Marco Chen <marcochen@chromium.org>
Reviewed-by: Pin-chih Lin <johnylin@chromium.org>