init: Change main.d to support running in background or foreground.
In CL:422588 goofy.sh was modified to run in background, but it seems
like we can't wait for "background process executed by a child process
which as also run in background". i.e., the 'wait' in
startup:factory_main does not work.

To solve that, we are changing main.d to always run in background, and a
special folder "pre-start" that rules inside will be executed in
foreground before anything in main.d was started.

By the change we can allow non-parallel rules (for example, setting
static IP) and parallel main programs to live in both main.d.

BUG=chromium:679676
TEST=status factory # see started/running.

Change-Id: I7f4068954a8834c9888558cd418c16f6f2dd7058
Reviewed-on: https://chromium-review.googlesource.com/428616
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>