setup: add util-linux to Dockerfile
The default `mount` command in Alpine does not support the "offset"
option, which is essential when importing bundles. However, the `mount`
command in util-linux supports it, so we'll install the package to solve
the problem.

The only drawback of installing the package is that the size of the
Docker image has grown 700KB, but since it's rather small so should be
acceptable.

BUG=b:34332927
TEST=By the following steps:
    1. Go into the Umpire Docker instance:
       `docker exec -it umpire_${BOARD} sh
    2. Import bundle:
       `umpire import-bundle --id ${NAME} --note ${NOTE} ${BUNDLE_PATH}`
    and saw it succeeded.

Change-Id: I9937cda4c017419dd3bd19bf58c9a23d138b08a3
Reviewed-on: https://chromium-review.googlesource.com/428529
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>