hwid builder: allow to add extra field into the pattern.
Originally we do not allow to add extra field into the pattern because
it would cause ambiguity. After adding a new kind of field into the
pattern, we will decode the old HWID string to index 0 for the field,
but we don't know if all old devices have the component what maps to
index 0.

But in practice, we face the scenario in early build:
There is a component that cannot probed after the first HWID config
file is created, and there is no second source for this component.
In this case, we can append the field into the pattern directly after
it can be probed.

BUG=chromium:678690
TEST=make test

Change-Id: Ib5d4826b1e3a0b31879b77e749f41c1a6fcff573
Reviewed-on: https://chromium-review.googlesource.com/429750
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>