hwid: Set default initial encoded fields.
We list the important components and preserve the encoded field for
them.

The first group of fields are always created. We add a default item if
they don't exist in the probe result.
[- image_id: 5]
- board_version_field: 3
- region_field: 5
- customization_id: 5
- cpu_field: 3
- storage_field: 5
- dram_field: 5

The second group of field are added in order if they exist in the
probe result.
- firmware_keys_field: 3
- ro_main_firmware: 3
- ro_ec_firmware: 2

BUG=chromium:678690
TEST=make test

Change-Id: Icc374ba33859719c3b839847fde86a81a37d7b17
Reviewed-on: https://chromium-review.googlesource.com/429970
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>