hwid: add use case in README.md.
We add the yaml file and the command for each use case, and use
gitiles syntax to add the table of content (TOC), and show the
difference of the yaml file by two-column layout.
Also we replace the word "transfer" by "convert" to describe the
action of replacing legacy style of region field.

BUG=chromium:678690
TEST=none

Change-Id: Ibca21e802b6ec7e0aa0b5a4011cf06719b408993
Reviewed-on: https://chromium-review.googlesource.com/431545
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>