instalog: output_archive: lower memory usage
Although creating archives in-memory could be considered a
"cleaner" and "safer" solution, problems were encountered
when creating large archives (~200mb).  Python memory usage
was skyrocketing to the point that the kernel would kill the
Instalog process.

Thus we modify output_archive to instead output to a directory
on disk.  Once the directory is fully prepared, we create an
archive based on its contents.

Also add a unittest, including a test for memory usage.

BUG=None
TEST=output_archive_unittest.py

Change-Id: I13354303d2379e0e68b817e9ac7847b6463091b4
Reviewed-on: https://chromium-review.googlesource.com/434557
Commit-Ready: Joel Kitching <kitching@chromium.org>
Tested-by: Joel Kitching <kitching@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>