Makefile: Add 'publish-docs' command.
The factory SDK document, especially region table, is very helpful for
partners and non-factory developers, even if they don't have built
bundles.

Particular for region table, currently we have it stored in several
places: a spreadsheet version in cros-keyboard-request, a web version in
cros-regions (but partners can't access it).

The new command 'publish-docs' tries to simplify the process so we have
a centralized place for delivering SDK documents in
 https://storage.googleapis.com/chromeos-factory-docs/sdk/index.html

BUG=None
TEST=make publish-docs

Change-Id: Ib3971040c626b8d3ded6f3f4ab7753659b6329f3
Reviewed-on: https://chromium-review.googlesource.com/439964
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>