testlog: update session once when logging series data.
Originally, testlog updates the session to JSON file when logging each
point. In VSWR pytest, we record the trace data which has 1600 points.
That means we need to update the session 1600 times, which takes lots
of time.
In this CL, we only update the session once after logging all the data
point. Thus, the test time is much shorter.

BUG=chromium:692401
TEST=make test

Change-Id: I3d44a19975a81b931755b615bd17468ed1df35ac
Reviewed-on: https://chromium-review.googlesource.com/443007
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>