goofy: implement goofy_server
Currently cros.factory.test.state is not only a state data storage, but
also a complicated server that:
   1. Serves static files for Goofy UI.
   2. Handles RPC for Goofy.
   3. Partially handles test list status.

To achieve this, we see code that inject goofy_rpc functions into state
instance to get RPC support, which is not really intuitive.

As we'd like to support RPC for Goofy plugins, state server will be too
complicated and hard to maintain.

In this CL, we implement a goofy_server that is designed to handle above
memtioned web request, including serving static files for UI and
handling multiple RPC instances.

The goal is to make the responsibility of each components more clear:
  - state: handles shared data.
  - goofy/goofy_rpc: organizes the system and handles test status.
  - goofy_server: handles all web requests.
  - goofy plugins: additional functions for goofy.

BUG=chromium:658151
TEST=make test
     run amd64-generic-goofy toolkit.

Change-Id: Ie5b50fef113114542f3574e75a625046f99d5063
Reviewed-on: https://chromium-review.googlesource.com/446142
Commit-Ready: Shun-Hsing Ou <shunhsingou@chromium.org>
Tested-by: Shun-Hsing Ou <shunhsingou@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>