device: Remove unused GetBatteryRegisters.
The 'GetBatteryRegisters' was only ever implemented on Spring. For most
platforms, it's not necessary to capture battery registers in factory.
Even on Spring, it was solely used for fault analysis.

As a result, this should be removed from general implementation
(device/power.py) and standard test (pytests/charger.py).

BUG=chromium:674459
TEST=make test

Change-Id: I33455e845046d8506ea647aeb761a6b34bc5843d
Reviewed-on: https://chromium-review.googlesource.com/453539
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>