whale: ScanBarcode support different file names
ScanBarcode in whale is actually "get saved barcode from NUC".
(The function name is inherited from previous projects.)

And there was a restriction to ScanBarcode that it always returns the
content of a pre-specified file.

This CL adds a second argument to the function, so we can get content
from a different file.

BUG=b:35985483
TEST=manual

Change-Id: Iaed4ba54c638e69e8948d0bee3a5fe3083549b31
Reviewed-on: https://chromium-review.googlesource.com/456720
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>