pytest summary: notify external test complete
Recently, we created a new test `wait_external_test`, which is still in
private overlay (CL:*314004).

The new argument 'run_factory_external_name' (which is the same with
`wait_external_test`) will make summary test create the file that
`wait_external_test` is waiting for.  The content of the file will be
'PASS' if all test in current test group is completed, otherwise, it
will be a list of path, status pairs, showing the result of each
subtests.

BUG=b:35985483
TEST=manual

Change-Id: I025f30701716308eff1e3a45e6a8fc038a4a6411
Reviewed-on: https://chromium-review.googlesource.com/456901
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>