goofy: Prepare DUT link after the Goofy plugins start.
Originally we verify the DUT link preparation when we initialize the
test list, which is executed before the Goofy plugins start. If the
connection manager plugin is enabled, and we use SSH link with DHCP
server, then the DHCP server would be broken because we restart shill
when the connection manager is enabled, which happened after the DHCP
server is created.

In this CL we move the execution of link preparation after starting
Goofy plugins.

BUG=b:36542462
TEST=manually start Goofy with SSH link and connection_manager

Change-Id: Ib9cb00858190f86c4b8cb6fe62157cac81a70037
Reviewed-on: https://chromium-review.googlesource.com/456981
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>