wifi_throughput: Support DHCP host IP.
In station-based testing, we execute Iperf server at the host
station. Originally we should setup a static IP for the host
station, and it's not flexible to duplicate the station.

Now we can assign a IP range by CIDR. Then we find the host's
IP which is in this range, and execute Iperf server at this IP.

BUG=none
TEST=manually test at device

Change-Id: I189c243ac0cb40cf148bbb7782161aeeb4a01a94
Reviewed-on: https://chromium-review.googlesource.com/458185
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>