dolphin fixture: re-raise socket error with BFTFixtureException
If the dolphin fixture is a remote device, bft_fixture will create a
socket connection to control the device.  If the connection cannot be
established, then a socket error will be raise.  This error is not
wrapped by BFTFixtureException, therefore, a usage like:

try:
  bft_fixture.CreateBFTFixture(**args)
except BFTFixtureException as e:
  # handle e ...

will failed to catch socket error.  In this CL, we re-raise the socket
error with a BFTFixtureException.

BUG=b:35985483
TEST=manual

Change-Id: I6da278d0a17664bd9bc0e0f647822421de7e0aec
Reviewed-on: https://chromium-review.googlesource.com/459159
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>