info: read Ethernet address file by ReadSpecialFile API.
When we read the Ethernet address file, which is at
"/sys/class/net/eth*/address", the "broken pipe" error occurs if
we use normal scp method. We changed to use ReadSpecialFile method
to read the content of the file.

Also, to prevent the bug, we try to use ReadSpecialFile if the
ReadFile fails.

BUG=none
TEST=manually test

Change-Id: I53ba86e14e6af423327c8fe506f6e339386eee5c
Reviewed-on: https://chromium-review.googlesource.com/459638
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>