Fix net_utils.FindUnusedTCPPort.
The code did not properly closed the TCPServer, and was relying on GC to
release the port before the returned port is actually used.

Actually starting a TCPServer is also unnecessary since we only need to
bind the port but not listen, and TCPServer would do both.

Listening only on LOCALHOST also cause problem sometimes, since some
code use the returned port to bind on 0.0.0.0.

Change this to open a socket, bind it on 0.0.0.0 port 0, and return the
bound port after closing it. There always would be race condition if the
port is used by others between the port is returned and actually used by
caller, but since almost all usage of this function is in unittest, and
the chance of race condition is very small, this can be ignored.

BUG=chromium:665306
TEST=make test

Change-Id: Ib67e2cd75d5405a0d8a84634ce5f3727a34dd094
Reviewed-on: https://chromium-review.googlesource.com/469531
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>