probe: fix probe_cmdline cannot evaluate shell function.
The argument name of `shell` function is 'command', which is
conflicted with the argument name used in probe_cmdline.py.
Now we changed the argument name in probe_cmdline.py to
`_Command`, which should not be conflicted with any arguments
name with lower_case_with_underline style.

BUG=none
TEST=./py/probe/probe_cmdline_unittest.py

Change-Id: Id3450cdaee2a4c601da054b666e16a8df36ca861
Reviewed-on: https://chromium-review.googlesource.com/471326
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>