raiden_charge: add option wait_after_engage_secs
We found that the charging / discharging current might not be stable
immediately after engaging charging power.
For example, in a test trial, the sampled ina current is:
[2980, -2, 0, -1, 2536, 2977, 2978, 2977].  And the average value does
not meet the spec.

In this CL, we add an option `wait_after_engage_secs`, which will make
pytest wait for a period of time before start sampling.
This option does not apply to engaging device for Check5VINACurrent,
because we want to make sure the current get into desired range in a
limited period of time.

BUG=None
TEST=None

Change-Id: I76f9011110d348162b230e1672ae5dfc57de2e56
Reviewed-on: https://chromium-review.googlesource.com/472406
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Wei-Han Chen <stimim@chromium.org>