wifi: retry to find access point.
In device wifi module, the function stack for finding AP is:
  FindAccessPoint(): Return the first matched AP object.
  FilterAccessPoints(): Return the list of matched AP objects.
  _AllAccessPoints(): Return the list of all found AP objects.

Originally we only added the retry mechanism at _AllAccessPoints to
handle the issue that Wifi device is busy sometimes. But it cannot
handle the situation that AP is unstable to scan.

Now we move the retry mechanism to FilterAccessPoints. If this
function doesn't find any matched AP, then it will retry until
it reaches the timeout limit.

BUG=none
TEST=manually test wifi_throughput pytest

Change-Id: Ifdad095124e3cb8fcec0b31aa39ca7cf6a82c845
Reviewed-on: https://chromium-review.googlesource.com/476532
Commit-Ready: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>