setup: Add 'pygpt.py' utility.
To simplify RMA image creation in future, we want to store resources on
its stateful partition and the most simple (and correct) way is to allow
enlarging RMA USB image.

To do that, we need an utility to correct GPT after an image file is
resized. Unfortunately, there's no existing widely available tool can do
that (cgpt is very close, but it has some bug and hard to bundle).

As a result, here's the new tool made from scratch and dedicated for
factory flow to use.

BUG=chromium:711615
TEST=truncate -s $((3*1048576*1024)) factory_install.bin
     cgpt show factory_install.bin
     pygpt repair --expand factory_install.bin
     pyget show factory_install.bin
     cgpt show factory_install.bin
     ./pygpt_unittest.py

Change-Id: I1d50def371404c4badf32b6f35e3947bf9b307a7
Reviewed-on: https://chromium-review.googlesource.com/477814
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>