test-list: implement expression checker
Use checker.AssertExpresionIsValid to check if an expression is allowed
to be used in the test list.

In new JSON test list, we allow users to write 'eval! ...' as an
expression that needs to be evaluated.  Since we want to prevent
arbitrary code evaluation, there are some constrain to the string.

1. the string will be compiled by: ast.parse(string, mode='eval').
   therefore, if the string is not a single expression, an exception
   will be raise.
2. the expression cannot be a lambda function or generator, since both
   of them are not serializable.
3. only builtin functions, `constants`, `options`, `dut`, `station`,
   `session` can be used in the expression.  if the expression contains
   other unbounded variables, an exception will be raised.

BUG=b:36206161
TEST=make test

Change-Id: Ia426510d8fb65a255e78bc5cdb38d4cc705f49e1
Reviewed-on: https://chromium-review.googlesource.com/479496
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>