i18n: Fix unicode handling problem.
There are places where the passed in dict to Translated would contains
text that are of type str, but are actually having content UTF-8
encoded (For example, label got from metadata in ui_template).

Since all i18n handling assume that the text is of type unicode, we
convert them to unicode if the value is str by UTF-8 decoding.

BUG=chromium:678434
TEST=make test
TEST=manually, run a test with label manually set to non-ascii character
     in test list, and see no exception.

Change-Id: Ide262758d4cbe0fc4b114ff20bf1095a45917729
Reviewed-on: https://chromium-review.googlesource.com/494367
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>