instalog: Properly close server in input_http.
The TCPServer accept socket needs to be closed by server_close(), or the
port would still be open after the request handling thread is finished.

This would cause about 1.5% of unittest failure because of timeout in
testShutdown, since if the post thread starts slow (after the
time.sleep(0.2)), request.post would still be able to connect to the
port after sandbox.Stop(), but would hang forever.

Also add a new unittest for this, and refactor some unittest to make it
more robust.

BUG=chromium:665306
TEST=make test

Change-Id: I44ab3d14a97ac214f44b9aef8b30116492ff0da5
Reviewed-on: https://chromium-review.googlesource.com/497239
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>