cros_payload: Add 'download' command for special resources.
To preserve content from release image stateful partition (for example,
pre-cached extensions) we want a new command to download arbitrary
resources (for instance, release_image.part1).

This is different from 'install' command because even if dest is a block
device, we still have to save it as file instead of overwriting the
block device.

BUG=chromium:711615
TEST=cros_payload download URL . release_image.part1

Change-Id: I1f2f40ea44b0e5191d92c598f6f40bddd608c6c6
Reviewed-on: https://chromium-review.googlesource.com/499811
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>