Speed up goofy_unittest.
The event_server.shutdown() has same issue as TCPServer.shutdown(), that
would take 0.5 seconds to complete. Change it to use
net_utils.ShutdownTCPServer, and also change net_utils.ShutdownTCPServer
to work on UnixStreamServer (which is a subclass of TCPServer).

This reduce the running time of goofy_unittest from about 18 seconds to
about 14 seconds, and the unittest is no longer the slowest one in `make
test`.

CQ-DEPEND=CL:503991
BUG=chromium:665306
TEST=make test

Change-Id: I966857b6a9915ec4a3d445dc7d2d410da9fcc60a
Reviewed-on: https://chromium-review.googlesource.com/503990
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>