Fix GetUnusedPort causing unittest flakyness.
The original GetUnusedPort use the "bind on port 0" technique to find an
unused port. Since the port got in this way would be in the ephemeral
port range, the approach has some drawback:

* The port would not only has race condition for other server calling
  GetUnusedPort, but it's also possible that the port to conflict with
  all outgoing connections.

* A typical usage of GetUnusedPort is as follows:

  port = net_utils.GetUnusedPort()
  server = StartServerAsync(port=port)
  sync_utils.WaitFor(lambda: PingServer(port=port), 2)

  But if the server port is in ephemeral port range, there's a small
  chance that the PingServer() would actually connects to itself,
  binding the port, and the StartServerAsync() would fail.
  (See http://sgros.blogspot.tw/2013/08/tcp-client-self-connect.html)

The most "correct" way for solving this would be enforcing that all
server act correctly when directly binding to port 0, and has some way
of reporting the bound port back to client, but that would be lots of
work.

A simple workaround is just randomly choose a port NOT in ephemeral port
range, and check if it's used. Since the ephemeral port range is >=
32768 on a typical Ubuntu machine, we choose 16384~32767.

This reduced the fail rate of "make test" by about 1%, from 25 fails in
1457 runs to 12 fails in 1457 runs.

Also refactor shopfloor_unittest a bit so it's less flaky.

BUG=chromium:665306
TEST=make test

Change-Id: Idf6feda8ce60698a467898e198016ba5f78d50b3
Reviewed-on: https://chromium-review.googlesource.com/503994
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>