download_patch: ignore failure and show more information
When we failed to cherry-pick a CL, the script would stop, thus all
following CLs will not be cherry-picked either.

In this CL, the script will recover from current failure and keep trying
the next CL.  And in the end, cherry-pick result for each change will be
printed.

BUG=b:35985269
TEST=manual

Change-Id: I401f0c72527d3463d8daf6ce4b5f6d3850445ac9
Reviewed-on: https://chromium-review.googlesource.com/536622
Commit-Ready: Wei-Han Chen <stimim@chromium.org>
Tested-by: Wei-Han Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>