removable_storage: Fix test stuck when timeout.
The self._ui.Fail would fail the pytest. If the FactoryTestFailure is
also raised, since it's called in the thread of CountdownTimer, which
has interrupt_on_crash set, the exception would cause a interrupt on the
main thread, and then the current failing pytest would be interrupted,
causing the test to be stuck in some weird state.

Fix by don't raise FactoryTestFailure, since self._ui.Fail is enough to
fail the test.

BUG=None
TEST=make test
TEST=manually, run the pytest on DUT, and wait it to timeout, check the
     pytest is properly marked as fail.

Change-Id: I67a205ac3bbca5ea1e6cd308710f6f61df1a250f
Reviewed-on: https://chromium-review.googlesource.com/505811
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Shun-Hsing Ou <shunhsingou@chromium.org>