Fix all unittest flakyness issues with GetUnusedPort.
After CL:503994, there's still about 0.5% chance that port got from
net_utils.GetUnusedPort would be the same for two different unittest by
chance, thus causing the test to fail with error "Address already in
use".

To completely solve this issue, we add a simple UnixStreamServer in
run_tests.py, that would manage current port range. When GetUnusedPort
sees that a special environment variable is set, it would try to connect
to that unix domain socket and acquire a random port that is guaranteed
not used by any other caller of GetUnusedPort.

The fail rates drops from about 1.1% (10 out of 919 runs) to about 0.55%
(15 out of 2743 runs), and there's zero test that fails with "Address
already in use" in these 15 fail runs.

Also renamed GetUnusedPort to FindUnusedPort, to be consistent to
FindUnusedTCPPort, and merged FindConsecutiveUnusedPorts to
FindUnusedPort, so the FindConsecutiveUnusedPorts can use the same
mechanic to avoid port conflict.

BUG=chromium:665306
TEST=make test
TEST=run "make test" many times, observe that there's no error with
     "Address already in use".

Change-Id: If08369b648190fad8cabe04e7d65c269601c0ec2
Reviewed-on: https://chromium-review.googlesource.com/508489
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>