Fix ext_display test always pass.
Pytests implements the FactoryTask differently regarding on how threads
are handled. Some FactoryTask calls Pass/Fail directly in Run (common
when there's no need of interaction), and some FactoryTask calls
Pass/Fail in the callback function (For example, ui.BindKey).

This cause problem since the next task to run is actually run in the
same thread as where the last Pass/Fail is called, so there are some
FactoryTask that would be run in the UI thread, which is bad since if
the next task is running directly in the UI thread, it would cause the
"Mark failed" in UI not working.

Also, some FactoryTask's Run() (e.g. ext_display) returns after spawning
a non-daemon thread, so the main thread exits, and rely on the fact that
the non-daemon thread would still be running.

Since CL:431331, the UI thread is always in the main thread. This cause
that the FactoryTaskManager.Run() returns immediately, and the
test_ui.RunInBackground thinks that the test had passed.

Refactor the FactoryTaskManager to solve all these problems, so the
FactoryTaskManager.Run() actually runs task in sequence in the same
thread, and wait for each task to finish (When the task call
Pass/Fail/Stop).

BUG=None
TEST=make test
TEST=manually, run ext_display test on DUT, and check that both ESC and
     "Mark failed" work as intended.

Change-Id: Ib93c97d137de3f3ef45fa116e06f7809dcc320a7
Reviewed-on: https://chromium-review.googlesource.com/509330
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>