pytests: Add pytest to update Cr50 firmware.
Add a pytest that can be used to force update the firmware of Cr50.

The firmware would come from either the release rootfs on DUT (default),
or from a specified path if dargs={'from_release': False,
'firmware_file': '...'} is set.

A RebootStep() should be added after this pytest to complete the update.

BUG=chromium:724526
TEST=make test

Change-Id: I24fe65e070e673f626c8895dae72499e107a10b6
Reviewed-on: https://chromium-review.googlesource.com/509602
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>