evdev_utils: Change FindDevice to use args as filters.
Change evdev_utils to use args as filters to filter things, and don't
stop at the first non-None.

For example, in the keyboard test, the args are 'name_fragment' and
evdev_utils.IsKeyboardDevice. It's possible that there are two devices
with exactly the same name, but only one of them is keyboard, it would
raise a 'not exactly one candidates' in the original code if we specify
name_fragment, but we still want to be able to specify which keyboard to
use.

BUG=None
TEST=make test
TEST=manually run keyboard pytest on DUT.

Change-Id: I23b39aec2ffffe18859951c89c310b285730cb15
Reviewed-on: https://chromium-review.googlesource.com/516783
Commit-Ready: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Youcheng Syu <youcheng@google.com>